requirements.txt
Django~=2.2.17
djangorestframework
django-cors-headers~=3.10.0
psycopg2-binary==2.8.6

